requirements.txt
django
django-ckeditor
pillow
django_mptt
whitenoise